Create version of Dragonfly without required engine-related code
This removes the engines, grammar, rpc, language, parsing sub-
packages, any engine-related test code and some files not relevant
to OS actions. Documentation files have been left in tact for now.

This branch is NOT to be merged with the main branch. It is intended
to be used as a starting point for the dictation-toolbox/dtactions
project.